refactor(service-client): move Item to under the components/atoms
https://github.com/akky-xxxx/project-my-homepage/issues/44
