+++ packages/service-client/src/components/atoms/Item/index.stories.tsx
@@ -1,4 +1,4 @@
-import { DeviceMap } from "../../../../../../../../../shared/const/DeviceMap"
+import { DeviceMap } from "../../../shared/const/DeviceMap"
 import { Item } from "./index"
 
 import type { ComponentStoryObj, Meta } from "@storybook/react"
